fix styles ( form_register ) : change styles  InputPhoneUserProfile
